Move extended implementation of telegram api to tdlib.py
